refactor(index): extract FAISS store layout
Move RFC 014 load/save and active index path logic out of FaissIndexManager while preserving its private wrappers and existing behavior.

Closes #133
